refactor(enrich/maintainer): split into submodule directory (#57 2/3)
src/enrich/maintainer.rs (1284 LOC) -> src/enrich/maintainer/:
- types.rs (74): Host, MaintainerAgeFinding, MaintainerInfo, LookupError, GithubResponse, GitlabResponse, constants
- pipeline.rs (206): enrich, enrich_with, enrich_with_hosts
- github.rs (118): lookup_github_repo, github_get
- gitlab.rs (129): lookup_gitlab_repo, gitlab_get
- codeberg.rs (19): lookup_codeberg_repo
- parsing.rs (286): URL/ISO/percent/JSON/Link-header helpers
- tests.rs (421): full #[cfg(test)] block

Biggest non-test file: parsing.rs at 286 LOC. Public API unchanged
(re-exported via mod.rs). Pure mechanical split, zero behavior change.
429 lib tests pass, clippy clean.
